Fix some sign/size conversion warnings from clang
psgeom.c:1320:17: warning: implicit conversion changes signedness:
 'int' to 'unsigned long' [-Wsign-conversion]
    if ((sym & (~0xff)) == 0x700) {
             ~  ^~~~~
psgeom.c:1666:31: warning: implicit conversion changes signedness:
 'int' to 'unsigned long' [-Wsign-conversion]
            else if (((sym & (~0xff)) == 0) && isprint(sym) && (!isspace(sym))) {
                           ~  ^~~~~
psgeom.c:1860:27: warning: implicit conversion loses integer precision:
 'size_t' (aka 'unsigned long') to 'int' [-Wshorten-64-to-32]
                int len = strlen(top->label[i]);
                    ~~~   ^~~~~~~~~~~~~~~~~~~~~
xkbprint.c:460:19: warning: implicit conversion loses integer precision:
 'size_t' (aka 'unsigned long') to 'int' [-Wshorten-64-to-32]
        int len = strlen(inputFile);
            ~~~   ^~~~~~~~~~~~~~~~~

+++ xkbprint.c
@@ -453,7 +453,7 @@ parseArgs(int argc, char *argv[])
         /* Nothing */
     }
     else if (strchr(inputFile, ':') == NULL) {
-        int len = strlen(inputFile);
+        size_t len = strlen(inputFile);
 
         if ((len > 4) && (strcmp(&inputFile[len - 4], ".xkm") == 0)) {
             /* Nothing */
